Update validation.php
Removed the Dutch translation for 'amount' since this is only one possible meaning and a specific one. This translation is 'the amount of money' or 'price'. A better translation for the general form/meaning of 'amount' would be 'hoeveelheid'. Since this word isn't it other language files (checked FR, DE and the English source), I recommend to remove it to avoid confusion.